Rename namespace to au.com.sealink.printing
