+++ src/main/java/au/com/sealink/printing/ticket_printer/exceptions/NoSuchPrinterException.java
@@ -1,0 +1,14 @@
+package au.com.sealink.printing.ticket_printer.exceptions;
+
+public class NoSuchPrinterException extends Exception {
+
+  private static final long serialVersionUID = 1L;
+
+  public NoSuchPrinterException() {
+    super();
+  }
+
+  public NoSuchPrinterException(String message) {
+    super(message);
+  }
+}
