+++ src/main/java/au/com/sealink/printing/ticket_printer/printables/PrintableElement.java
@@ -1,0 +1,21 @@
+package au.com.sealink.printing.ticket_printer.printables;
+
+import java.awt.Graphics2D;
+
+import au.com.sealink.printing.ticket_printer.TicketElement;
+
+public abstract class PrintableElement {
+
+  TicketElement element;
+
+  // Fonts seem to need this multiplier to accurately position... 
+  protected final double fontMultiplier = 0.8;
+
+  public PrintableElement(TicketElement element) {
+    this.element = element;
+  }
+
+  // Each printable element type implements this
+  public abstract void drawOn(Graphics2D g);
+
+}
