--- src/main/java/com/quicktravel/ticket_printer/exceptions/NoSuchPrinterException.java
@@ -1,14 +1,0 @@
-package com.quicktravel.ticket_printer.exceptions;
-
-public class NoSuchPrinterException extends Exception {
-
-  private static final long serialVersionUID = 1L;
-
-  public NoSuchPrinterException() {
-    super();
-  }
-
-  public NoSuchPrinterException(String message) {
-    super(message);
-  }
-}
