--- src/main/java/com/quicktravel/ticket_printer/exceptions/NoTicketPageSettingsAssigned.java
@@ -1,15 +1,0 @@
-package com.quicktravel.ticket_printer.exceptions;
-
-public class NoTicketPageSettingsAssigned extends Exception {
-
-  private static final long serialVersionUID = 1L;
-
-  public NoTicketPageSettingsAssigned() {
-    super();
-  }
-
-  public NoTicketPageSettingsAssigned(String message) {
-    super(message);
-  }
-
-}
