--- src/main/java/com/quicktravel/ticket_printer/printables/PrintableNullElement.java
@@ -1,18 +1,0 @@
-package com.quicktravel.ticket_printer.printables;
-
-import java.awt.Graphics2D;
-
-import com.quicktravel.ticket_printer.TicketElement;
-
-class PrintableNullElement extends PrintableElement {
-
-  public PrintableNullElement(TicketElement element) {
-    super(element);
-  }
-
-  @Override
-  public void drawOn(Graphics2D g) {
-    // null object
-  }
-
-}
